fix: switch to vitest
