--- frontend-vite/babel.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-    presets: [
-      '@babel/preset-env',
-      ['@babel/preset-react', {runtime: 'automatic'}],
-    ],
-  };
