+++ frontend-vite/package.json
@@ -9,7 +9,7 @@
     "lint": "eslint .",
     "preview": "vite preview",
     "postinstall": "flowbite-react patch",
-    "test": "jest"
+    "test": "vitest"
   },
   "dependencies": {
     "@sentry/react": "^9.9.0",
@@ -27,23 +27,22 @@
     "tailwindcss": "^4.0.17"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.26.9",
-    "@babel/preset-react": "^7.26.3",
     "@eslint/js": "^9.23.0",
     "@testing-library/dom": "^10.4.0",
+    "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
+    "@testing-library/user-event": "^14.6.1",
     "@types/react": "^19.0.12",
     "@types/react-dom": "^19.0.4",
     "@vitejs/plugin-react": "^4.3.4",
-    "babel-jest": "^29.7.0",
     "eslint": "^9.23.0",
     "eslint-plugin-react-compiler": "^19.0.0-beta-aeaed83-20250323",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.19",
     "globals": "^16.0.0",
-    "jest": "^29.7.0",
-    "react-test-renderer": "^19.0.0",
-    "vite": "^6.2.3"
+    "jsdom": "^26.0.0",
+    "vite": "^6.2.3",
+    "vitest": "^3.0.9"
   },
   "packageManager": "yarn@3.8.7+sha512.bbe7e310ff7fd20dc63b111110f96fe18192234bb0d4f10441fa6b85d2b644c8923db8fbe6d7886257ace948440ab1f83325ad02af457a1806cdc97f03d2508e"
 }
